refactor: restructure Button and Section components
- Updated Button component to use default export.
- Introduced a new Section component for better layout management.
- Cleaned up imports in the main page file to reflect the new structure.
